Move survey step files
